refactor(core): complete nio migration and remove dual-stack architecture
This commit completes the migration from `CocoaAsyncSocket` to `SwiftNIO` by:

- DELETE: `BaseConnection.swift` (dual-stack abstraction layer)
- DELETE: `NIOConnection.swift` (legacy filename, now canonical `Connection`)
- DELETE: `NIODownloadTask.swift` (legacy filename, now canonical `DownloadTask`)
- DELETE: `NIOUploadTask.swift` (legacy filename, now canonical `UploadTask`)

With NIO now the sole implementation, all "NIO" prefixes are redundant:

Core types:
  - `NIOConnection` → `Connection`
  - `NIODownloadTask` → `DownloadTask`
  - `NIOUploadTask` → `UploadTask`
  - `NIOConnectionDelegate` → `ConnectionDelegate` (protocol)
  - `NIOConnectionError` → `ConnectionError` (enum)
  - `NIOPayloadClientTrustHandler` → `PayloadClientTrustHandler`
  - `NIOPayloadTrustHandler` → `PayloadTrustHandler`
  - `NIODownloadHandler` → `DownloadHandler`
  - `NIOUploadHandler` → `UploadHandler`
  - `NIOTrustHandler` → `TrustHandler`
  - `NIOCertificateUtils` → `SSLCertificateUtils`
  - `NIOUdpHandler` → `UdpHandler`
  - `NIOTcpConnectionHandler` → `TcpConnectionHandler`

ConnectionProvider (removed NIO-prefixed property names):
  - `nioEventLoopGroup` → `eventLoopGroup`
  - `nioUdpChannel` → `udpChannel`
  - `nioTcpServerChannel` → `tcpServerChannel`
  - `nioTcpListeningPort` → `tcpListeningPort`
  - `startNIOUdp()` → `startUdp()`
  - `stopNIOUdp()` → `stopUdp()`
  - `startNIOTcpServer()` → `startTcpServer()`
  - `stopNIOTcpServer()` → `stopTcpServer()`
  - `convertNIOAddressToSocketAddress()` → `convertToSocketAddress()`

Throughout codebase, simplified variable names since types are now canonical:
  - `nioAddress` → `targetAddress` / `address`
  - `nioSSLCert` → `sslCert`
  - `nioSSLKey` → `sslKey`
  - `deviceNIOAddress` → `deviceAddress`
  - `firstNIOCert` → `peerCert`
  - `nioCertificate` → `sslCertificate`
  - `convertedNioCert` → `convertedCert`

- Removed "legacy" terminology from all comments and docstrings
- Removed references to CocoaAsyncSocket/GCDAsyncSocket (no longer relevant)
- Simplified class documentation (removed "NIO-based", "NIO-specific" language)
- Updated comments to reflect current architecture without dual-stack context:
  - "Convert legacy SocketAddress to NIOCore.SocketAddress"
    → "Convert to NIOCore.SocketAddress"
  - "Extract host IP from legacy SocketAddress" → "Extract host IP from peer address"
  - "TCP keepalive configuration matching the existing CocoaAsyncSocket settings"
    → "TCP keepalive configuration for network connections"
  - Removed outdated references to methods that no longer exist

All 13 services consolidated to single `Connection`/`DownloadTask`/`UploadTask`:
  - Removed dual download task delegate methods (`NIODownloadTaskDelegate` → `DownloadTaskDelegate`)
  - Removed alternative protocol handlers
  - Updated all service implementations to use canonical types
  - Fixed logging calls that referenced NIO task types directly

This starts the removal of `CocoaAsyncSocket` dependency. The codebase is now
pure SwiftNIO with no dual-stack complexity or redundant naming conventions.
