+++ Soduto/Core/ConnectionProvider.swift
@@ -8,21 +8,10 @@
 
 import Foundation
 import Cocoa
-import CocoaAsyncSocket
-import os
 import Network
 import NIOCore
 import NIOPosix
-
-// MARK: - Feature Flags
-
-/// Feature flag to enable NIO TCP server implementation.
-/// Set to `true` to use SwiftNIO ServerBootstrap instead of GCDAsyncSocket for accepting connections.
-private let USE_NIO_TCP_SERVER = true
-
-/// Feature flag to enable NIO for outgoing connections.
-/// Set to `true` to use NIOConnection with ClientBootstrap for connections initiated by us.
-private let USE_NIO_OUTGOING = true
+import os
 
 enum ConnectionProviderError: Error {
     case IdentityAbsent
@@ -31,10 +20,9 @@ enum ConnectionProviderError: Error {
 public protocol ConnectionProviderDelegate: AnyObject {
     func isNewConnectionNeeded(byProvider provider: ConnectionProvider, deviceId: String) -> Bool
     func connectionProvider(_ provider: ConnectionProvider, didCreateConnection: Connection)
-    func connectionProvider(_ provider: ConnectionProvider, didCreateNIOConnection: NIOConnection)
 }
 
-public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDelegate, NIOConnectionDelegate {
+public class ConnectionProvider: NSObject, ConnectionDelegate {
     
     static public let udpPort: UInt16 = 1716
     static public let minTcpPort: UInt16 = 1716
@@ -49,26 +37,24 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
     private let config: ConnectionConfiguration
     private let pathMonitor: NWPathMonitor = NWPathMonitor()
     private let pathMonitorQueue: DispatchQueue = DispatchQueue(label: "com.soduto.NetworkMonitor")
-    private let tcpSocket: GCDAsyncSocket = GCDAsyncSocket(delegate: nil, delegateQueue: DispatchQueue.main)
     private var pendingConnections: Set<Connection> = Set<Connection>()
-    private var pendingNIOConnections: Set<NIOConnection> = Set<NIOConnection>()
     private var isStarted: Bool = false
     private var lastAnnouncementTime: TimeInterval = 0.0
     private var announcementTimer: Timer? = nil
     
-    // MARK: NIO Properties
+    // MARK: Network Properties
     
-    /// NIO event loop group for network operations.
-    private var nioEventLoopGroup: MultiThreadedEventLoopGroup?
+    /// Event loop group for network operations.
+    private var eventLoopGroup: MultiThreadedEventLoopGroup?
     
-    /// NIO UDP channel for receiving broadcasts.
-    private var nioUdpChannel: Channel?
+    /// UDP channel for receiving broadcasts.
+    private var udpChannel: Channel?
     
-    /// NIO TCP server channel for accepting connections.
-    private var nioTcpServerChannel: Channel?
+    /// TCP server channel for accepting connections.
+    private var tcpServerChannel: Channel?
     
-    /// The port the NIO TCP server is listening on.
-    private var nioTcpListeningPort: UInt16 = 0
+    /// The port the TCP server is listening on.
+    private var tcpListeningPort: UInt16 = 0
     
     
     
@@ -86,7 +72,6 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
                 }
             }
         }
-        self.tcpSocket.delegate = self
         
         NotificationCenter.default.addObserver(self, selector: #selector(broadcastAnnouncement), name: ConnectionProvider.broadcastAnnouncementNotification, object: nil)
     }
@@ -96,29 +81,16 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         NotificationCenter.default.removeObserver(self)
         
         // Shutdown NIO event loop group
-        try? self.nioEventLoopGroup?.syncShutdownGracefully()
+        try? self.eventLoopGroup?.syncShutdownGracefully()
     }
     
     public func start() {
         
         // Listen for device announcement broadcasts
-        startNIOUdp()
+        startUdp()
         
         // Listen for connections on TCP
-        if USE_NIO_TCP_SERVER {
-            startNIOTcpServer()
-        } else {
-            for port: UInt16 in ConnectionProvider.minTcpPort...ConnectionProvider.maxTcpPort {
-                do {
-                    try self.tcpSocket.accept(onPort: port)
-                    Logger.network.info("Listening for TCP connections on port \(self.tcpSocket.localPort, privacy: .public)")
-                }
-                catch {}
-            }
-            if self.tcpSocket.isDisconnected {
-                Logger.network.error("Failed to start listening TCP connections on ports in range \(ConnectionProvider.minTcpPort, privacy: .public)-\(ConnectionProvider.maxTcpPort, privacy: .public)")
-            }
-        }
+        startTcpServer()
         
         self.isStarted = true
         
@@ -127,7 +99,6 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         broadcastAnnouncement()
         
         // Speculative broadcasts after some intervals.
-        // When broadcasting imediately after internet connection becomes available, ARP table may be incomplete and not all known devices may be detected. After some time, theese undetected devices may become known and may receive the announcement
         _ = Timer.compatScheduledTimer(withTimeInterval: 40.0, repeats: false) { _ in self.broadcastAnnouncement() }
         _ = Timer.compatScheduledTimer(withTimeInterval: 80.0, repeats: false) { _ in self.broadcastAnnouncement() }
         _ = Timer.compatScheduledTimer(withTimeInterval: 120.0, repeats: false) { _ in self.broadcastAnnouncement() }
@@ -136,12 +107,8 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
     public func stop() {
         self.isStarted = false
         self.pathMonitor.cancel()
-        stopNIOUdp()
-        if USE_NIO_TCP_SERVER {
-            stopNIOTcpServer()
-        } else {
-            self.tcpSocket.disconnect()
-        }
+        stopUdp()
+        stopTcpServer()
     }
     
     public func restart() {
@@ -156,15 +123,8 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
     @objc public dynamic func broadcastAnnouncement() {
         guard self.isStarted else { return }
         
-        // Get the TCP port from the appropriate server
-        let tcpPort: UInt16
-        if USE_NIO_TCP_SERVER {
-            guard self.nioTcpListeningPort > 0 else { return }
-            tcpPort = self.nioTcpListeningPort
-        } else {
-            guard self.tcpSocket.localPort > 0 else { return }
-            tcpPort = self.tcpSocket.localPort
-        }
+        guard self.tcpListeningPort > 0 else { return }
+        let tcpPort = self.tcpListeningPort
         
         guard self.announcementTimer == nil else { return }
         
@@ -192,7 +152,7 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
     
     /// Sends the identity packet as a UDP broadcast.
     private func sendBroadcast(packet: DataPacket) {
-        guard let channel = self.nioUdpChannel else {
+        guard let channel = self.udpChannel else {
             Logger.network.error("UDP channel not available for broadcast")
             return
         }
@@ -220,34 +180,16 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
             for deviceConfig in knownDeviceConfigs {
                 guard deviceConfig.hwAddresses.contains(accessibleHwAddress) else { continue }
                 do {
-                    let deviceNIOAddress = try NIOCore.SocketAddress(ipAddress: accessibleAddress.ipAddressString, port: Int(ConnectionProvider.udpPort))
+                    let deviceAddress = try NIOCore.SocketAddress(ipAddress: accessibleAddress.ipAddressString, port: Int(ConnectionProvider.udpPort))
                     var deviceBuffer = channel.allocator.buffer(capacity: bytes.count)
                     deviceBuffer.writeBytes(bytes)
-                    let envelope = AddressedEnvelope(remoteAddress: deviceNIOAddress, data: deviceBuffer)
+                    let envelope = AddressedEnvelope(remoteAddress: deviceAddress, data: deviceBuffer)
                     channel.writeAndFlush(envelope, promise: nil)
                 } catch {
                     Logger.network.error("Failed to send to known device: \(error, privacy: .public)")
                 }
                 break
             }
-        }
-    }
-    
-    // MARK: GCDAsyncSocketDelegate
-    
-    public func newSocketQueueForConnection(fromAddress address: Data, on sock: GCDAsyncSocket) -> DispatchQueue? {
-        return DispatchQueue.main
-    }
-    
-    public func socket(_ sock: GCDAsyncSocket, didAcceptNewSocket newSocket: GCDAsyncSocket) {
-        Logger.network.debug("socket(<\(sock, privacy: .public)> didAcceptNewSocket:<\(newSocket, privacy: .public)>)")
-        
-        if let connection = Connection(socket: newSocket, config: self.config) {
-            connection.delegate = self
-            self.pendingConnections.insert(connection)
-            
-            // read initial identity packet
-            connection.readOnePacket()
         }
     }
     
@@ -264,19 +206,19 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
                 connection.readPackets()
                 self.pendingConnections.remove(connection)
                 delegate.connectionProvider(self, didCreateConnection: connection)
-            }
-            else {
-                Logger.network.error("No connection provider delegate to take new connection - closing");
+            } else {
+                Logger.network.error("No connection provider delegate to take new connection - closing")
                 connection.close()
             }
         default:
-            assert(false, "Closed or Open connection state expected")
+            break
         }
     }
     
     public func connection(_ connection: Connection, didSendPacket packet: DataPacket, uploadedPayload: Bool) {
         Logger.network.debug("connection(<\(connection, privacy: .public)> didSendPacket:<\(packet, privacy: .public)>)")
         
+        // After sending identity packet (outgoing connection), secure as server
         do {
             guard let identity = connection.identity else { throw ConnectionProviderError.IdentityAbsent }
             let protocolVersion = try identity.getProtocolVersion()
@@ -293,7 +235,12 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
     }
     
     public func connection(_ connection: Connection, didReadPacket packet: DataPacket) {
-        Logger.network.debug("connection(<\(connection, privacy: .public)> didReadPacket:<\(packet, privacy: .public)>)")
+        Logger.network.debug("connection(<\(connection, privacy: .public)> didReadPacket:<\(packet.type, privacy: .public)>)")
+        
+        // Only process identity packet during initialization
+        guard connection.state == .Initializing else {
+            return
+        }
         
         // The only packet we are waiting for is first identity packet to initialize connection with
         do {
@@ -314,86 +261,16 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
     public func connectionCapacityChanged(_ connection: Connection) { }
     
     
-    // MARK: NIOConnectionDelegate
+    // MARK: - UDP Implementation
     
-    public func nioConnection(_ connection: NIOConnection, didSwitchToState state: NIOConnection.State) {
-        Logger.network.debug("connection(<\(String(describing: connection), privacy: .public)> switchedToState:<\(String(describing: state), privacy: .public)>)")
-        switch state {
-        case .Closed:
-            self.pendingNIOConnections.remove(connection)
-        case .Open:
-            if let delegate = self.delegate {
-                connection.readPackets()
-                self.pendingNIOConnections.remove(connection)
-                delegate.connectionProvider(self, didCreateNIOConnection: connection)
-            } else {
-                Logger.network.error("No connection provider delegate to take new connection - closing")
-                connection.close()
-            }
-        default:
-            break
-        }
-    }
-    
-    public func nioConnection(_ connection: NIOConnection, didSendPacket packet: DataPacket, uploadedPayload: Bool) {
-        Logger.network.debug("connection(<\(connection, privacy: .public)> didSendPacket:<\(packet, privacy: .public)>)")
-        
-        // After sending identity packet (outgoing connection), secure as server
-        do {
-            guard let identity = connection.identity else { throw ConnectionProviderError.IdentityAbsent }
-            let protocolVersion = try identity.getProtocolVersion()
-            if protocolVersion >= ConnectionProvider.minVersionWithSSLSupport {
-                // Beware that securing as server while connection initiated by self
-                connection.secureServer()
-            }
-            connection.finishInitialization()
-        }
-        catch {
-            Logger.network.error("Failed to initialize connection: \(error, privacy: .public)")
-            connection.close()
-        }
-    }
-    
-    public func nioConnection(_ connection: NIOConnection, didReadPacket packet: DataPacket) {
-        Logger.network.debug("connection(<\(connection, privacy: .public)> didReadPacket:<\(packet.type, privacy: .public)>)")
-        
-        // Only process identity packet during initialization
-        guard connection.state == .Initializing else {
-            // Connection already initialized - packet would normally go to Device
-            // For now, just log it since Device integration is pending
-            Logger.network.debug("Received \(packet.type, privacy: .public) packet on open connection (Device integration pending)")
-            return
-        }
-        
-        // The only packet we are waiting for is first identity packet to initialize connection with
-        do {
-            try connection.applyIdentity(packet: packet)
-            let protocolVersion = try packet.getProtocolVersion()
-            if protocolVersion >= ConnectionProvider.minVersionWithSSLSupport {
-                // Beware that securing as client while connection initiated by the peer
-                connection.secureClient()
-            }
-            connection.finishInitialization()
-        }
-        catch {
-            Logger.network.error("Failed to initialize connection: \(error, privacy: .public)")
-            connection.close()
-        }
-    }
-    
-    public func nioConnectionCapacityChanged(_ connection: NIOConnection) { }
-    
-    
-    // MARK: - NIO UDP Implementation
-    
-    /// Starts the NIO UDP channel for receiving and sending broadcasts.
-    private func startNIOUdp() {
+    /// Starts the UDP channel for receiving and sending broadcasts.
+    private func startUdp() {
         // Create event loop group if needed
-        if self.nioEventLoopGroup == nil {
-            self.nioEventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
+        if self.eventLoopGroup == nil {
+            self.eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
         }
         
-        guard let group = self.nioEventLoopGroup else {
+        guard let group = self.eventLoopGroup else {
             Logger.network.error("Failed to create event loop group")
             return
         }
@@ -406,26 +283,26 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
                 guard let self = self else {
                     return channel.eventLoop.makeSucceededVoidFuture()
                 }
-                let handler = NIOUdpHandler(connectionProvider: self, config: self.config)
+                let handler = UdpHandler(connectionProvider: self, config: self.config)
                 return channel.pipeline.addHandler(handler)
             }
         
         do {
             let channel = try bootstrap.bind(host: "0.0.0.0", port: Int(ConnectionProvider.udpPort)).wait()
-            self.nioUdpChannel = channel
+            self.udpChannel = channel
             Logger.network.info("Listening for UDP broadcasts on port \(ConnectionProvider.udpPort, privacy: .public)")
         } catch {
             Logger.network.error("Failed to start UDP: \(error, privacy: .public)")
         }
     }
     
-    /// Stops the NIO UDP channel.
-    private func stopNIOUdp() {
-        self.nioUdpChannel?.close(promise: nil)
-        self.nioUdpChannel = nil
+    /// Stops the UDP channel.
+    private func stopUdp() {
+        self.udpChannel?.close(promise: nil)
+        self.udpChannel = nil
     }
     
-    /// Handles incoming UDP packet from NIO.
+    /// Handles incoming UDP packet.
     fileprivate func handleUdpPacket(data: Data, remoteAddress: NIOCore.SocketAddress) {
         guard let delegate = self.delegate else { return }
         guard let packet = DataPacket(data: data) else { return }
@@ -440,7 +317,7 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
 #endif
         
         // Create a socket address for the connection
-        guard let connectionAddress = convertNIOAddressToSocketAddress(remoteAddress, port: UInt16(port)) else {
+        guard let connectionAddress = convertToSocketAddress(remoteAddress, port: UInt16(port)) else {
             Logger.network.error("Failed to convert address to SocketAddress")
             return
         }
@@ -448,34 +325,21 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         // Dispatch to main queue for connection creation
         DispatchQueue.main.async { [weak self] in
             guard let self = self else { return }
-            
-            if USE_NIO_OUTGOING {
-                // Use NIOConnection for outgoing connections
-                self.createNIOOutgoingConnection(address: connectionAddress, identityPacket: packet)
-            } else {
-                // Use legacy Connection
-                if let connection = Connection(address: connectionAddress, identityPacket: packet, config: self.config) {
-                    connection.delegate = self
-                    self.pendingConnections.insert(connection)
-                    
-                    // send initial identity packet
-                    _ = connection.send(DataPacket.identityPacket(config: self.config))
-                }
-            }
+            self.createOutgoingConnection(address: connectionAddress, identityPacket: packet)
         }
     }
     
     /// Creates an outgoing connection to the specified address.
-    private func createNIOOutgoingConnection(address: SocketAddress, identityPacket packet: DataPacket) {
-        guard let group = self.nioEventLoopGroup else {
+    private func createOutgoingConnection(address: SocketAddress, identityPacket packet: DataPacket) {
+        guard let group = self.eventLoopGroup else {
             Logger.network.error("No event loop group available for outgoing connection")
             return
         }
         
-        if let connection = NIOConnection(address: address, identityPacket: packet, config: self.config, eventLoopGroup: group) {
+        if let connection = Connection(address: address, identityPacket: packet, config: self.config, eventLoopGroup: group) {
             Logger.network.debug("Created outgoing connection to \(address.description, privacy: .public)")
             connection.delegate = self
-            self.pendingNIOConnections.insert(connection)
+            self.pendingConnections.insert(connection)
             
             // Send initial identity packet
             _ = connection.send(DataPacket.identityPacket(config: self.config))
@@ -484,9 +348,9 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         }
     }
     
-    /// Converts a NIO SocketAddress to the legacy SocketAddress type.
-    private func convertNIOAddressToSocketAddress(_ nioAddress: NIOCore.SocketAddress, port: UInt16) -> SocketAddress? {
-        switch nioAddress {
+    /// Converts a NIOCore.SocketAddress to SocketAddress.
+    private func convertToSocketAddress(_ address: NIOCore.SocketAddress, port: UInt16) -> SocketAddress? {
+        switch address {
         case .v4(let addr):
             var socketAddress = SocketAddress(addr: addr.address)
             socketAddress.port = in_port_t(port)
@@ -500,16 +364,16 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         }
     }
     
-    // MARK: - NIO TCP Server Implementation
+    // MARK: - TCP Server Implementation
     
-    /// Starts the NIO TCP server for accepting incoming connections.
-    private func startNIOTcpServer() {
+    /// Starts the TCP server for accepting incoming connections.
+    private func startTcpServer() {
         // Create event loop group if needed (shared with UDP)
-        if self.nioEventLoopGroup == nil {
-            self.nioEventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
+        if self.eventLoopGroup == nil {
+            self.eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
         }
         
-        guard let group = self.nioEventLoopGroup else {
+        guard let group = self.eventLoopGroup else {
             Logger.network.error("Failed to create event loop group for TCP server")
             return
         }
@@ -522,9 +386,7 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
                 guard let self = self else {
                     return channel.eventLoop.makeSucceededVoidFuture()
                 }
-                // For now, just add a handler that notifies us of new connections
-                // Full connection handling will be implemented with NIOConnection
-                let handler = NIOTcpConnectionHandler(connectionProvider: self, config: self.config)
+                let handler = TcpConnectionHandler(connectionProvider: self, config: self.config)
                 return channel.pipeline.addHandler(handler)
             }
         
@@ -532,8 +394,8 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         for port in ConnectionProvider.minTcpPort...ConnectionProvider.maxTcpPort {
             do {
                 let channel = try bootstrap.bind(host: "0.0.0.0", port: Int(port)).wait()
-                self.nioTcpServerChannel = channel
-                self.nioTcpListeningPort = port
+                self.tcpServerChannel = channel
+                self.tcpListeningPort = port
                 Logger.network.info("Listening for TCP connections on port \(port, privacy: .public)")
                 return
             } catch {
@@ -545,27 +407,27 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         Logger.network.error("Failed to start TCP server on ports \(ConnectionProvider.minTcpPort, privacy: .public)-\(ConnectionProvider.maxTcpPort, privacy: .public)")
     }
     
-    /// Stops the NIO TCP server.
-    private func stopNIOTcpServer() {
-        self.nioTcpServerChannel?.close(promise: nil)
-        self.nioTcpServerChannel = nil
-        self.nioTcpListeningPort = 0
+    /// Stops the TCP server.
+    private func stopTcpServer() {
+        self.tcpServerChannel?.close(promise: nil)
+        self.tcpServerChannel = nil
+        self.tcpListeningPort = 0
     }
     
-    /// Handles a new TCP connection accepted by the NIO server.
+    /// Handles a new TCP connection accepted by the server.
     fileprivate func handleTcpConnection(channel: Channel, remoteAddress: NIOCore.SocketAddress) {
         Logger.network.debug("TCP accepted connection from \(String(describing: remoteAddress), privacy: .public)")
         
-        guard let group = self.nioEventLoopGroup else {
+        guard let group = self.eventLoopGroup else {
             Logger.network.error("No event loop group available for connection")
             channel.close(promise: nil)
             return
         }
         
         // Create connection from the accepted channel
-        if let connection = NIOConnection(channel: channel, config: self.config, eventLoopGroup: group) {
+        if let connection = Connection(channel: channel, config: self.config, eventLoopGroup: group) {
             connection.delegate = self
-            self.pendingNIOConnections.insert(connection)
+            self.pendingConnections.insert(connection)
             
             // Read initial identity packet
             connection.readOnePacket()
@@ -575,7 +437,7 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
         }
     }
     
-    // MARK: Private methrod
+    // MARK: Private methods
     
     private func becameReachable() {
         Logger.network.debug("Became reachable")
@@ -589,10 +451,10 @@ public class ConnectionProvider: NSObject, GCDAsyncSocketDelegate, ConnectionDel
     
 }
 
-// MARK: - NIO UDP Handler
+// MARK: - UDP Handler
 
-/// Channel handler for receiving UDP datagrams via SwiftNIO.
-private final class NIOUdpHandler: ChannelInboundHandler {
+/// Channel handler for receiving UDP datagrams.
+private final class UdpHandler: ChannelInboundHandler {
     typealias InboundIn = AddressedEnvelope<ByteBuffer>
     
     private weak var connectionProvider: ConnectionProvider?
@@ -621,12 +483,12 @@ private final class NIOUdpHandler: ChannelInboundHandler {
     }
 }
 
-// MARK: - NIO TCP Connection Handler
+// MARK: - TCP Connection Handler
 
-/// Channel handler for TCP connections accepted by the NIO server.
+/// Channel handler for TCP connections accepted by the server.
 /// This handler simply notifies ConnectionProvider and then removes itself,
-/// allowing NIOConnection to take over the channel pipeline.
-private final class NIOTcpConnectionHandler: ChannelInboundHandler, RemovableChannelHandler {
+/// allowing Connection to take over the channel pipeline.
+private final class TcpConnectionHandler: ChannelInboundHandler, RemovableChannelHandler {
     typealias InboundIn = ByteBuffer
     
     private weak var connectionProvider: ConnectionProvider?
@@ -643,7 +505,7 @@ private final class NIOTcpConnectionHandler: ChannelInboundHandler, RemovableCha
             connectionProvider?.handleTcpConnection(channel: context.channel, remoteAddress: remoteAddress)
         }
         
-        // Remove ourselves from the pipeline - NIOConnection will add its own handlers
+        // Remove ourselves from the pipeline - Connection will add its own handlers
         context.pipeline.removeHandler(self, promise: nil)
     }
