+++ Soduto/Core/DataPacket.swift
@@ -37,7 +37,6 @@ public struct DataPacket: CustomStringConvertible {
     var payloadSize: Int64? = nil
     var payloadInfo: PayloadInfo?
     var downloadTask: DownloadTask? = nil
-    var nioDownloadTask: NIODownloadTask? = nil
     
     public var description: String {
         do {
@@ -118,7 +117,7 @@ public struct DataPacket: CustomStringConvertible {
     }
     
     public func hasPayload() -> Bool {
-        return self.payload != nil || self.downloadTask != nil || self.nioDownloadTask != nil
+        return self.payload != nil || self.downloadTask != nil
     }
