+++ Soduto/Core/Pairable.swift
@@ -16,17 +16,14 @@ public enum PairingStatus: Int {
 }
 
 public struct PairingRequest {
-    
-    let connection: Connection
-    
+    public let connection: Connection
 }
 
+/// Delegate protocol for Connection pairing events.
 public protocol PairableDelegate: AnyObject {
-    
-    func pairable(_ pairable:Pairable, receivedRequest request:PairingRequest)
-    func pairable(_ pairable:Pairable, failedWithError error:Error)
-    func pairable(_ pairable:Pairable, statusChanged status:PairingStatus)
-    
+    func connection(_ connection: Connection, receivedPairingRequest request: PairingRequest)
+    func connection(_ connection: Connection, pairingFailed error: Error)
+    func connection(_ connection: Connection, pairingStatusChanged status: PairingStatus)
 }
 
 public protocol Pairable {
