+++ Soduto/Services/TelephonyService.swift
@@ -64,16 +64,7 @@ public class TelephonyService: Service, UserNotificationActionHandler {
     
     // MARK: Service methods
     
-    /// NIO-compatible packet handler.
-    public func handleDataPacket(_ dataPacket: DataPacket, fromDevice device: Device, onAnyConnection connection: AnyBaseConnection) -> Bool {
-        return handleDataPacketCore(dataPacket, fromDevice: device)
-    }
-    
     public func handleDataPacket(_ dataPacket: DataPacket, fromDevice device: Device, onConnection connection: Connection) -> Bool {
-        return handleDataPacketCore(dataPacket, fromDevice: device)
-    }
-    
-    private func handleDataPacketCore(_ dataPacket: DataPacket, fromDevice device: Device) -> Bool {
         guard dataPacket.isTelephonyPacket else { return false }
         
 #if DEBUG
@@ -109,7 +100,7 @@ public class TelephonyService: Service, UserNotificationActionHandler {
         catch {
             Logger.services.error("Error while handling telephony packet: \(error, privacy: .public)")
         }
-
+        
         return true
     }
